+++ relay-filter/src/common.rs
@@ -127,6 +127,9 @@ pub enum FilterStatKey {
 
     /// Filtered due to invalid CSP policy.
     InvalidCsp,
+
+    /// Filtered due to the fact that it was a call to a filtered transaction
+    FilteredTransactions,
 }
 
 // An event grouped to a removed group.
@@ -153,6 +156,7 @@ impl FilterStatKey {
             FilterStatKey::Localhost => "localhost",
             FilterStatKey::WebCrawlers => "web-crawlers",
             FilterStatKey::InvalidCsp => "invalid-csp",
+            FilterStatKey::FilteredTransactions => "filtered-transaction",
         }
     }
 }
@@ -176,6 +180,7 @@ impl<'a> TryFrom<&'a str> for FilterStatKey {
             "localhost" => FilterStatKey::Localhost,
             "web-crawlers" => FilterStatKey::WebCrawlers,
             "invalid-csp" => FilterStatKey::InvalidCsp,
+            "filtered-transaction" => FilterStatKey::FilteredTransactions,
             other => {
                 return Err(other);
             }