+++ py/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add filter based on transaction names. ([#2118](https://github.com/getsentry/relay/pull/2118))
 - Add `lock` attribute to the frame protocol. ([#2171](https://github.com/getsentry/relay/pull/2171))
 
 ## 0.8.25